requirements.txt
# asgiref==3.4.1
certifi==2021.10.8

charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.8

django-cors-headers==3.10.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.2
